Commands System refactor to be used in ESPCam
